Updated PHPDoc documentation.
git-svn-id: http://lucene-silverstripe-plugin.googlecode.com/svn/trunk@10 37c722ee-a693-3632-ac79-18467f915a76

+++ code/ZendSearchLuceneWrapper.php
@@ -1,31 +1,50 @@
 <?php
 
-/*******************************************************************************
- * Provides indexing and searching facilities.
- *******************************************************************************/
+/**
+ * Provides a wrapper to Zend Search Lucene.
+ *
+ * @package lucene-silverstripe-plugin
+ */
 
 class ZendSearchLuceneWrapper {
 
-    /* The name of the index. */
+    /**
+     * The name of the index. 
+     * @static
+     */
     public static $indexName = 'Silverstripe';
 
-    /* Stores a handle to the search index so we don't have to keep recreating it */
+    /** 
+     * Stores a handle to the search index so we don't have to keep recreating it. 
+     * @static
+     */
     private static $index = false;
 
     /**
      * Returns a set of results from Zend Search Lucene from the given search
      * parameters.
      *
-	 * @param Mixed $query
-	 * @return DataObjectSet
+     * It is possible to either pass in a string (which is what the default 
+     * form provided by this package does), or build your own query using the 
+     * Zend_Search_Lucene query building API.
+     *
+     * @link http://zendframework.com/manual/en/zend.search.lucene.searching.html
+     *
+	 * @param   Mixed           $query  String or object to pass to the find() method
+	 *                                  of the index.
+     * @link http://framework.zend.com/apidoc/core/Zend_Search_Lucene/Zend_Search_Lucene_Proxy.html#find
+	 * @return  Array           An array of Zend_Search_Lucene_Search_QueryHit 
+	 *                          objects representing the results of the search.
      */
     public static function find($query) {
-        $index = self::getIndex();        
+        $index = self::getIndex();
+        echo get_class( $index );
         try {
             $hits = $index->find($query);
         } catch ( Exception $e) {
             $hits = array();
         }
+        var_dump( $hits);
 		return $hits;
     }
 
@@ -61,8 +80,16 @@ class ZendSearchLuceneWrapper {
 
     /**
      * Indexes a DataObject.
+     *
+     * @param   DataObject  $object     The DataObject to index.  If the DataObject
+     *                                  does not have the ZendSearchLuceneSearchable
+     *                                  extension, it is not indexed.
      */
     public static function index($object) {
+        if ( ! Object::has_extension($object->ClassName, 'ZendSearchLuceneSearchable') ) {
+            return;
+        }
+
         $index = self::getIndex();
 
         // Remove currently indexed data for this object
@@ -89,14 +116,21 @@ class ZendSearchLuceneWrapper {
 
     /**
      * Delete a DataObject from the search index.
+     *
+     * @param   DataObject  $object     The DataObject to delete from the index.  If the DataObject
+     *                                  does not have the ZendSearchLuceneSearchable
+     *                                  extension, it is not deleted.
      */
     public static function delete($object) {
+        if ( ! Object::has_extension($object->ClassName, 'ZendSearchLuceneSearchable') ) {
+            return;
+        }
+
         $index = self::getIndex();
         $term = new Zend_Search_Lucene_Index_Term($object->ID, 'ID');
         foreach ($index->termDocs($term) as $id) {
             $doc = $index->getDocument($id);
             if ( $doc->ClassName != $object->ClassName ) continue;
-$out .= 'existing record found, deleted: ';
             $index->delete($id);
         }
     }
@@ -109,13 +143,21 @@ $out .= 'existing record found, deleted: ';
      * Keyword - Data that is searchable and stored in the index, but not 
      *      broken up into tokens for indexing. This is useful for being 
      *      able to search on non-textual data such as IDs or URLs.
+     *
      * UnIndexed – Data that isn’t available for searching, but is stored 
      *      with our document (eg. article teaser, article URL  and timestamp 
      *      of creation)
+     *
      * UnStored – Data that is available for search, but isn’t stored in 
      *      the index in full (eg. the document content)
+     *
      * Text – Data that is available for search and is stored in full 
      *      (eg. title and author)
+     *
+     * @param   DataObject  $object     The DataObject from which to extract a
+     *                                  Zend field.
+     * @param   String      $fieldName  The name of the field to fetch a Zend field for.
+     * @return  Zend_Search_Lucene_Field
      */
     private static function getZendField($object, $fieldName) {
         $extraSearchFields = $object->getExtraSearchFields();
